snapshot reorg (#954)
