+++ canary/snapshotCollector.go
@@ -4,35 +4,10 @@ import (
 	"context"
 	"fmt"
 	influxdb2 "github.com/influxdata/influxdb-client-go/v2"
-	"github.com/openziti/zrok/util"
 	"github.com/sirupsen/logrus"
 	"slices"
 	"sort"
-	"time"
 )
-
-type Snapshot struct {
-	Operation string
-	Instance  uint
-	Iteration uint64
-	Started   time.Time
-	Completed time.Time
-	Ok        bool
-	Error     error
-	Size      uint64
-}
-
-func NewSnapshot(operation string, instance uint, iteration uint64) *Snapshot {
-	return &Snapshot{Operation: operation, Instance: instance, Iteration: iteration, Started: time.Now()}
-}
-
-func (s *Snapshot) String() string {
-	if s.Ok {
-		return fmt.Sprintf("[%v, %d, %d] (ok) %v, %v", s.Operation, s.Instance, s.Iteration, s.Completed.Sub(s.Started), util.BytesToSize(int64(s.Size)))
-	} else {
-		return fmt.Sprintf("[%v, %d, %d] (err) %v, %v, (%v)", s.Operation, s.Instance, s.Iteration, s.Completed.Sub(s.Started), util.BytesToSize(int64(s.Size)), s.Error)
-	}
-}
 
 type SnapshotCollector struct {
 	InputQueue chan *Snapshot
